Formatted code and added Kafka listening

+++ cart-service/utils/kafka_consume.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/segmentio/kafka-go"
 	log "github.com/sirupsen/logrus"
+	"github.com/swiggy-ipp/cart-service/configs"
 	"github.com/swiggy-ipp/cart-service/dto/requests"
 	"github.com/swiggy-ipp/cart-service/services"
 )
@@ -12,25 +13,27 @@ import (
 // KafkaCartConsumeService is a Kafka Consumer for Cart Service
 type KafkaCartConsumeService struct {
 	brokerAddress string
-	topic string
-	cartService services.CartService
+	topic         string
+	cartService   services.CartService
 }
 
 // Create a new Kafka Cart Consumer
 func NewKafkaCartConsumeService(
-	brokerAddress string, 
-	topic string, 
+	brokerAddress string,
+	topic string,
 	cartService services.CartService,
 ) *KafkaCartConsumeService {
 	return &KafkaCartConsumeService{
 		brokerAddress: brokerAddress,
-		topic: topic,
-		cartService: cartService,
+		topic:         topic,
+		cartService:   cartService,
 	}
 }
 
-// Consume the Kafka Cart Service and delete the cart of the user when user ID received on delete topic.
-func (kc *KafkaCartConsumeService) KafkaDeleteConsume() {
+// Consume the Kafka User IDs from Profile service.
+// Create the cart of the user when user ID received on Created topic.
+// Delete the cart of the user when user ID received on Deleted topic.
+func (kc *KafkaCartConsumeService) KafkaUserIDConsume() {
 	// Set up Background context for Kafka listener
 	ctx := context.Background()
 	l := log.New()
@@ -38,7 +41,7 @@ func (kc *KafkaCartConsumeService) KafkaDeleteConsume() {
 		Brokers: []string{kc.brokerAddress},
 		Topic:   kc.topic,
 		GroupID: "my-group",
-		Logger: l,
+		Logger:  l,
 	})
 
 	for {
@@ -51,10 +54,18 @@ func (kc *KafkaCartConsumeService) KafkaDeleteConsume() {
 		// After receiving the message, log its value and apply the callback
 		log.Info(string(msg.Key) + ": " + string(msg.Value))
 
-		// Delete Cart by User ID
-		err = kc.cartService.DeleteCart(ctx, requests.CartIDRequest{UserID: string(msg.Key)})
-		if err != nil {
-			log.Error("Could not delete cart: " + err.Error())
+		if kc.topic == configs.EnvKafkaUserCreatedTopic() {
+			// Create Cart by User ID
+			err = kc.cartService.CreateCart(ctx, string(msg.Value))
+			if err != nil {
+				log.Error("Could not create cart: " + err.Error())
+			}
+		} else if kc.topic == configs.EnvKafkaUserDeletedTopic() {
+			// Delete Cart by User ID
+			err = kc.cartService.DeleteCart(ctx, requests.CartIDRequest{UserID: string(msg.Key)})
+			if err != nil {
+				log.Error("Could not delete cart: " + err.Error())
+			}
 		}
 	}
 }
